+++ tests/Pim/Product/ProductServiceTest.php
@@ -7,9 +7,12 @@ declare(strict_types=1);
 
 namespace Wizaplace\SDK\Tests\Pim\Product;
 
+use Composer\Autoload\ClassLoader;
+use Doctrine\Common\Annotations\AnnotationRegistry;
 use GuzzleHttp\Psr7\Uri;
 use Psr\Http\Message\UriInterface;
 use Wizaplace\SDK\Exception\NotFound;
+use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
 use Wizaplace\SDK\Pagination;
 use Wizaplace\SDK\Pim\Product\CreateProductCommand;
 use Wizaplace\SDK\Pim\Product\Product;
@@ -25,7 +28,9 @@ use Wizaplace\SDK\Pim\Product\ProductListFilter;
 use Wizaplace\SDK\Pim\Product\ProductService;
 use Wizaplace\SDK\Pim\Product\ProductStatus;
 use Wizaplace\SDK\Pim\Product\ProductSummary;
+use Wizaplace\SDK\Pim\Product\Shipping;
 use Wizaplace\SDK\Pim\Product\UpdateProductCommand;
+use Wizaplace\SDK\Pim\Product\UpdateShippingCommand;
 use Wizaplace\SDK\Tests\ApiTestCase;
 
 final class ProductServiceTest extends ApiTestCase
@@ -812,11 +817,88 @@ final class ProductServiceTest extends ApiTestCase
         $this->assertNull($declinations[3]->getCode());
     }
 
+    public function testGetProductShipping()
+    {
+        $shipping = $this->buildProductService()->getShipping(5, 1);
+
+        $this->assertInstanceOf(Shipping::class, $shipping);
+    }
+
+    public function testGetProductShippings()
+    {
+        $shippings = $this->buildProductService()->getShippings(5);
+
+        foreach ($shippings as $shipping) {
+            $this->assertInstanceOf(Shipping::class, $shipping);
+        }
+    }
+
+    public function testPutProductShipping()
+    {
+        $this->loadAnnotations();
+
+        $command = new UpdateShippingCommand();
+        $command->setStatus("D")
+                ->setRates([
+                    [
+                        'amount' => 0,
+                        'value'  => 100,
+                    ],
+                    [
+                        'amount' => 1,
+                        'value'  => 50,
+                    ],
+                ])
+                ->setSpecificRate(false)
+                ->setProductId(5);
+
+        $productService = $this->buildProductService();
+
+        $productService->putShipping(1, $command);
+
+        $shipping = $productService->getShipping(5, 1);
+
+        $this->assertInstanceOf(Shipping::class, $shipping);
+        $this->assertSame(100.0, $shipping->getRates()[0]['value']);
+    }
+
+    public function testUpdateShippingCommandConstraints()
+    {
+        $this->loadAnnotations();
+
+        $command = new UpdateShippingCommand();
+        $command->setStatus("Status qui n'existe pas")
+            ->setRates([
+                [
+                    'amount' => 0,
+                    'value'  => 100,
+                ],
+                [
+                    'amount' => 1,
+                    'value'  => 50,
+                ],
+            ])
+            ->setSpecificRate(false)
+            ->setProductId(-1);
+
+        $productService = $this->buildProductService();
+
+        $this->expectException(SomeParametersAreInvalid::class);
+        $productService->putShipping(1, $command);
+    }
+
     private function buildProductService($userEmail = 'admin@wizaplace.com', $userPassword = 'password'): ProductService
     {
         $apiClient = $this->buildApiClient();
         $apiClient->authenticate($userEmail, $userPassword);
 
         return new ProductService($apiClient);
+    }
+
+    private function loadAnnotations() : void
+    {
+        /** @var ClassLoader $loader */
+        $loader = require __DIR__.'/../../../vendor/autoload.php';
+        AnnotationRegistry::registerLoader([$loader, 'loadClass']);
     }
 }