+++ CHANGELOG.md
@@ -1,14 +1,15 @@
 # Changelog
 
-## Next release
+## 1.40.0
 
-- Added `getUserId` to `\Wizaplace\SDK\Organisation\OrganisationBasket`
-- Added `\Wizaplace\SDK\Catalog\Product::getSeoKeywords`
-- Fix `seoDescription` from `Wizaplace\SDK\Catalog\Product` for MVP
+### New features
+
 - Added optional parameter `$language` to `\Wizaplace\SDK\Catalog\CatalogService::getAllProducts`
 - Added `\Wizaplace\SDK\Catalog\Shipping::getDeliveryTime`
 - Added `\Wizaplace\SDK\Basket\Shipping::getImage`
 - Added `\Wizaplace\SDK\Organisation\OrganisationService::checkoutBasket`
+- Updated `\Wizaplace\SDK\Organisation\OrganisationService::getOrganisationOrders`
+- Updated `\Wizaplace\SDK\Vendor\OrderSymmary`
 - Added `\Wizaplace\SDK\Catalog\Product::hasInfiniteStock`
 - Added `\Wizaplace\SDK\Catalog\Declination::hasInfiniteStock`
 - Added `\Wizaplace\SDK\Catalog\DeclinationSummary::hasInfiniteStock`
@@ -16,6 +17,28 @@
 - Added `\Wizaplace\SDK\PIM\Product\ProductDeclinationUpsertData::setInfiniteStock`
 - Added `\Wizaplace\SDK\PIM\Product\ProductUpsertData::setInfiniteStock`
 - Added `\Wizaplace\SDK\Order\OrderItem::getProductImageId`
+- Added `\Wizaplace\SDK\Vendor\Order\Order::getCompanyId`
+- Added `\Wizaplace\SDK\Vendor\Order\OrderSummary::getCompanyId`
+
+## 1.39.0
+
+### New features
+
+- Added `\Wizaplace\SDK\Pim\Product::getShippings`
+
+## 1.38.0
+
+### New features
+
+- Added `\Wizaplace\SDK\Organisation\OrganisationBasket::getUserId`
+- Added `\Wizaplace\SDK\Catalog\Product::getSeoKeywords`
+- Added `\Wizaplace\SDK\PIM\Product::getShipping`
+- Added `\Wizaplace\SDK\PIM\Product::putShipping`
+- Added `\Wizaplace\SDK\PIM\UpdateShippingCommand`
+
+### Corrections
+
+- Fixed an issue for MVP with `Wizaplace\SDK\Catalog\Product::getSeoDescription`
 
 ### 1.37.0